% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident}
\alias{accident}
\title{Accidents de la route aux États-Unis}
\format{
Une base de données de 40 observations en format long contenant les 5 variables suivantes:
\describe{
\item{annee}{facteur décrivant l'année de référence, 2010 ou 2018}
\item{region}{facteur indicant la région géographique telles que définies par la NHTSA}
\item{popn}{entier, population totale estimée de la région}
\item{nmorts}{entier, nombre de morts}
\item{moment}{facteur indiquant quelle période le décompte couvre, soit jour ou nuit}
}
}
\source{
\emph{National Highway Traffic Safety Administration} (NHTSA)
}
\usage{
accident
}
\description{
Accidents de la route aux États-Unis
}
\keyword{datasets}
