% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{affpol}
\alias{affpol}
\title{Affiliation politique d'Américain(e)s}
\format{
Un tableau de contingence deux par trois:
\describe{
\item{nombre}{décompte}
\item{sexe}{sexe, soit \code{homme} ou \code{femme}}
\item{parti}{facteur indiquant le parti, un parmi Démocrate (\code{democrate}), indépendant (\code{indep}) et GOP (\code{republicain})}
}
}
\source{
Tableau 2.5 de Agresti (2007), An Introduction to Categorical Data Analysis, Wiley.
}
\usage{
affpol
}
\description{
Données tirées du \emph{General Social Survey} (2000) qui recense
le sexe et l'identification politique des sondés selon le parti vers lequel
ils s'identifiaient le plus
}
\keyword{datasets}
