% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assurance}
\alias{assurance}
\title{Primes d'assurance santé}
\format{
Une base de données de 1338 observations avec
\describe{
  \item{age}{âge (en années)}
  \item{sexe}{sexe, \code{homme} ou \code{femme}}
  \item{imc}{indice de masse corporelle (en kg/mètres carrés)}
  \item{enfant}{nombre d'enfants à charge}
  \item{fumeur}{\code{oui} pour les fumeurs, \code{non} autrement.}
  \item{region}{lieu de résidence, une région parmi \code{sudouest}, \code{sudest}, \code{nordouest} ou \code{nordest}}
  \item{frais}{frais médicaux annuels en 2003 (en dollars USD).}
}
}
\source{
Lantz, Brett (2003), \emph{Machine Learning with R}, Packt Publishing.
}
\usage{
assurance
}
\description{
Le fichier de données simulées \code{assurance} contient de l'information sur les frais médicaux
facturés à 1338 adultes américains assurés au courant de l'année 2003.
}
\keyword{datasets}
