% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attente}
\alias{attente}
\title{Temps d'attente pour le métro}
\format{
un vecteur numérique de 62 observations
\describe{
\item{\code{attente}}{temps (en secondes) avant le départ du prochain métro}
}
}
\source{
STM
}
\usage{
attente
}
\description{
Temps (en secondes) de 17h59 jusqu'au départ de la prochaine rame de métro à la station Université de Montréal durant les jours de semaines. Les données ont été collectées pendant trois mois consécutifs.
}
\keyword{datasets}
