% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{automobile}
\alias{automobile}
\title{Caractéristiques d'automobiles}
\format{
une base de données de 392 observations contenant les variables suivantes
\describe{
\item{\code{autonomie}}{autonomie d'essence (en miles par gallons américains)}
\item{\code{cylindre}}{facteur, nombre de cylindres}
\item{\code{cylindree}}{cylindrée, le  volume balayé par les cylindres du moteur (en pouces cubiques)}
\item{\code{puissance}}{puissance du moteur (en chevaux-vapeurs)}
\item{\code{masse}}{masse de la voiture (en livres)}
\item{\code{acceleration}}{nombre de secondes pour que la voiture accélère de 0 à 60 miles à l'heure}
}
}
\source{
Ernesto Ramos et David Donoho; les données sont extraites de StatLib hébergie par l'Université Carnegie.
}
\usage{
automobile
}
\description{
Consommation d'essence, puissance, cylindrée et autres caractéristiques d'automobiles. Ces données ont servi à la foir de la Société américaine de statistique (ASA) en 1983.
}
\references{
James, G., Witten, D., Hastie, T., and Tibshirani, R. (2013) \emph{An Introduction to Statistical Learning with applications in R}, Springer-Verlag, New York
}
\keyword{datasets}
