% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bixi}
\alias{bixi}
\title{Emprunt de Bixis}
\format{
une base de donnée de 500 observations quotidiennes
\describe{
\item{\code{nutilisateurs}}{nombre de locations quotidiennes}
\item{\code{date}}{date }
\item{\code{temp}}{température à l'aéroport (en degrés Celcius)}
\item{\code{humid}}{pourcentage d'humidité relative}
\item{\code{jour}}{jour de la semaine, allant de dimanche (\code{1}) à samedi (\code{7})}
\item{\code{fds}}{indicateur binaire pour la fin de semaine (\code{0}), valant \code{1} en semaine}
}
}
\source{
Bixi, Environnement Canada (aéroport de Dorval)
}
\usage{
bixi
}
\description{
Ce jeu de données recense le nombre d'utilisateurs du système de location de
vélos libre-service Bixi à la station Édouard-Montpetit en fonction
du jour de l'année et de la météo (la température et le taux d'humidité relative)
}
\author{
Léo Belzile
}
\keyword{datasets}
