% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bixicol}
\alias{bixicol}
\title{Illustration de la colinéarité à l'aide des locations de Bixi}
\format{
une base de données de 1184 observations contenant les variables suivantes
\describe{
\item{\code{lognutilisateur}}{log du nombre total de locations quotidienne}
\item{\code{celcius}}{température moyenne à l'aéroport (en degrés Celcius)}
\item{\code{farenheit}}{température moyenne à l'aéroport (en degrés Farenheit)}
\item{\code{rfarenheit}}{température moyenne à l'aéroport (in Farenheit), arrondie au degré près}
}
}
\source{
Bixi, Environnement Canada
}
\usage{
bixicol
}
\description{
Ce jeu de données recense le nombre de locations quotidiennes de Bixi entre 2014 et 2019
en fonction de la température
de l'aéroport voisin de Dorval, enregistrée à 16h. Il sert à démontrer l'impact de la colinéarité.
}
\author{
Léo Belzile
}
\keyword{datasets}
