% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eolienne}
\alias{eolienne}
\title{Production électrique d'éoliennes}
\format{
une base de données de 25 observations des variables suivantes
\describe{
\item{\code{vitesse}}{vélocité des vents (en miles à l'heure)}
\item{\code{production}}{production électrique (en volts CC)}
}
}
\source{
Joglekar et al. (1989), Lack of Fit Testing when Replicates Are Not Available, \emph{The American Statistician}, 43,(\bold{3}), 135-143.
}
\usage{
eolienne
}
\description{
Mesures de production électrique d'une éolienne sur 25 plages distinctes de 15 minutes. 25 separate fifteen minute periods. We are interested in the relation between direct output
L'intérêt principal des données est la relation entre la production électrique et la vitesse moyenne des vents.
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}. Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\seealso{
\link[BSDA]{Windmill}
}
\keyword{datasets}
