% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbconduite}
\alias{gbconduite}
\title{Examens pratique de conduite}
\format{
Une base de données contenant les résultats suivants, regroupées par centre:
\describe{
 \item{reussite}{nombre de personnes ayant réussi l'examen (par sexe)}
 \item{total}{nombre total de personne dans le centre (par sexe)}
 \item{region}{facteur indiquant une des neuf régions d'Angleterre, le Pays de Galles ou l'Écosse (les données sont agrégées pour ces deux dernières)}
 \item{sexe}{facteur, \code{homme} ou \code{femme}}
 }
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/driving-test-statistics-drt}
}
\usage{
gbconduite
}
\description{
Ces données, tirées de GovUK et nettoyées, donnent le taux de succès
aux examens pratique de conduite en Grande-Bretagne pour 2018 dans 346 centres de conduite.
Un total de 761 750 personnes on réussi l’examen pratique, pour 1 663 897 essais.
}
\keyword{datasets}
