% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intention}
\alias{intention}
\title{Intention d'achat pour une publicité de bonbons}
\format{
une base de donnée de 120 observations contenant les variables suivantes
\describe{
\item{\code{intention}}{ variable discrète entre \code{2} et \code{14}; plus elle est élevée, plus le sujet exprime l’intention d’acheter ce produit.  Le score a été construit en additionnant les réponses de deux questions sur une échelle de Likert allant de fortement en désaccord (\code{1}) à fortement en accord (\code{7})}
\item{\code{fixation}}{durée totale de fixation de la publicité (en secondes).}
\item{\code{emotion}}{une mesure de la valence durant la fixation, soit le ratio de la probabilité d’une émotion positive sur la probabilité d’une émotion négative}
\item{\code{sexe}}{sexe du sujet, soit homme (\code{0}) ou femme (\code{1}).}
\item{\code{age}}{âge (en années).}
\item{\code{revenu}}{variable catégorique indiquant le revenu annuel du sujet; un parmi (\code{1}) [0, 20 000]; (\code{2}) [20 000, 60 000] ou (\code{3}) 60 000 et plus.}
\item{\code{educ}}{variable catégorique indiquant le niveau d’éducation le plus élevé obtenu, soit (\code{1}) secondaire ou moindre; (\code{2}) collégial, ou (\code{3}) universitaire.}
\item{\code{statut}}{statut matrimonial, soit célibaire (\code{0}) ou en couple (\code{1}).}
\item{\code{achat}}{variable binaire indiquant si la personne a acheté des bonbons suite au visionnement, oui (\code{1}) ou non (\code{0}).}
\item{\code{nachat}}{nombre de paquets de bonbons achetés}
}
}
\source{
Marc Fredette
}
\usage{
intention
}
\description{
Ces données simulées reproduisent une étude réalisée au Tech3Lab, dans le cadre de laquelle des cobayes devaient naviguer sur un site internet qui contenait, entre autres choses, une publicité pour des bonbons.
Pendant la navigation, un oculomètre mesurait l'endroit où se posait le regard du sujet.
On a ainsi pu mesurer si le sujet a regardé la publicité et combien de temps il l’a regardé.
De plus, un logiciel d’analyse des expressions faciales (FaceReader) a également été utilisé pour mesurer l’émotion du sujet pendant qu’il regardait la publicité.
À la fin de l’expérience, un questionnaire mesurait l’intention d’achat du sujet pour ces bonbons, ainsi que des variables socio-démographiques.
Seuls les 120 sujets qui ont regardé la publicité sont inclus dans les données.
}
\keyword{datasets}
