% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmat.R
\name{plot.lme_rescov}
\alias{plot.lme_rescov}
\title{Graphique de la matrice de covariance des Y}
\usage{
\method{plot}{lme_rescov}(x, corr = FALSE, ...)
}
\arguments{
\item{corr}{logique; si \code{TRUE}, le graphique représente la matrice de corrélation}

\item{...}{paramètres additionnels, actuellement ignorés}

\item{model}{objet de classe \code{hec_rescov}}
}
\description{
Graphique de la matrice de covariance des Y
}
