% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{renfe_simu}
\alias{renfe_simu}
\title{Résultats d'une étude de simulation sur les données Renfe}
\format{
Une base de données de 1000 observations
\describe{
\item{difmoy}{différence de moyenne de prix entre Madrid-Barcelone versus Barcelone-Madrid}
\item{statW}{statistique de Welch pour le test de localisation pour deux échantillons}
\item{icbi}{borne inférieure de l'intervalle de confiance basé sur la statistique de Welch}
\item{icbs}{borne supérieure de l'intervalle de confiance basé sur la statistique de Welch}
\item{valp}{valeur-p asymptotique}
}
}
\source{
Léo Belzile
}
\usage{
renfe_simu
}
\description{
Statistiques de test, valeur-p et différence de moyenne pour 1000
échantillons aléatoires de taille 100 tirées de la base de données
Renfe.
}
\keyword{datasets}
