% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmat.R
\name{rescov}
\alias{rescov}
\title{Matrice de covariance du vecteur réponse}
\source{
https://stackoverflow.com/questions/45650548/get-residual-variance-covariance-matrix-in-lme4/45655597#45655597
}
\usage{
rescov(model)
}
\arguments{
\item{model}{un modèle linéaire mixte, ou objet de classe \code{lmerMod}}
}
\description{
Matrice de covariance du vecteur réponse
}
\author{
Ben Bolker
}
