% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trafficaerien}
\alias{trafficaerien}
\title{Traffic aérien international}
\format{
une base de données de 144 observations contenant les variables suivantes
\describe{
\item{\code{passagers}}{nombre de passagers aériens internationaux (en milliers).}
\item{\code{mois}}{entier indiquant le mois, de janvier (\code{1}) à décembre (\code{12}).}
\item{\code{annee}}{année}
}
}
\source{
Box, G. E. P., Jenkins, G. M. and Reinsel, G. C. (1976) \emph{Time Series Analysis, Forecasting and Control}. Third Edition. Holden-Day. Series G.
}
\usage{
trafficaerien
}
\description{
Nombre mensuel total de passagers aériens internationaux entre 1949 et 1960.
}
\seealso{
\link[datasets]{AirPassengers}
}
\keyword{datasets}
