% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distraction}
\alias{distraction}
\title{Distraction face à un obstacle}
\format{
Une base de données de 35 observations
\describe{
\item{c}{temps de réaction en secondes lors d'une conversation}
\item{c}{temps de réaction en secondes lors d'une conversation}
}
}
\source{
Marc Fredette
}
\usage{
distraction
}
\description{
Données appariées fictives relative à une étude du Tech3Lab
comparant le temps de réaction (en secondes) de participants pour détecter
l'apparition d'un bicyclette dans leur champ visuel
exposés à deux mises en situations présentées dans un ordre aléatoire: texter
ou parler en marchant au cellulaire.
}
\keyword{datasets}
